proenDE2025-main/requirements.txt
# direct
argon2-cffi==23.1.0
fastapi==0.115.14
ipykernel==6.29.5
mysql-connector-python==9.4.0
ffmpeg-python==0.2.0
pwdlib==0.2.1
PyJWT==2.10.1
roboflow==1.2.1
scikit-learn==1.7.1
shapely==2.1.2
SQLAlchemy==2.0.41
ultralytics==8.3.160
uvicorn==0.34.3

# transitive
python-multipart==0.0.20
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.2
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
idna==3.7
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.1.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.13
